Upgrade LLVM to `18` for eBPF code compile
